Major refactor of code base.
