+++ pycollocation/__init__.py
@@ -2,14 +2,16 @@
 Objects imported here will live in the `pycollocation` namespace
 
 """
-__all__ = ["BoundaryValueProblem", "OrthogonalPolynomialSolver", "Solution"]
+__all__ = ["SymbolicBoundaryValueProblem", "OrthogonalPolynomialSolver", "Solution"]
 
 
+from . import boundary_value_problems
+from . import differential_equations
 from . import models
 from . import orthogonal_polynomials
 from . import solutions
 
-from . models import BoundaryValueProblem
+from . boundary_value_problems import SymbolicBoundaryValueProblem
 from . orthogonal_polynomials import OrthogonalPolynomialSolver
 from . solutions import Solution
